bug:auto错误,add无法找到文件夹“

+++ app/models/records.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, VARCHAR
+from sqlalchemy import Column, Integer, VARCHAR,FLOAT
 
 from app.db import Base
 from app.db.session import engine
@@ -14,7 +14,7 @@ class Records(Base):
     videoID = Column(Integer, comment="用户观看的作品编号")
     episodeID = Column(Integer, comment="用户观看的视频编号")
     src = Column(VARCHAR(128), comment="用户观看的视频src")
-    timeline = Column(Integer, comment="上次观看的时间轴")
+    timeline = Column(FLOAT, comment="上次观看的时间轴")
 
 
 Base.metadata.create_all(bind=engine)
